+++ src/main/java/com/project/library/controller/SubscriptionController.java
@@ -2,6 +2,9 @@ package com.project.library.controller;
 
 import com.project.library.Db.SubscriptionRepository;
 import com.project.library.model.Subscription;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
@@ -11,30 +14,68 @@ import java.util.List;
 @RequestMapping("/subscription")
 public class SubscriptionController {
 
+    Logger logger = LoggerFactory.getLogger(getClass());
+
     @Autowired
     private SubscriptionRepository subscriptionRepository;
 
     @GetMapping("/list")
     public List<Subscription> getList() {
-        return (List<Subscription>)subscriptionRepository.findAll();
+
+        List<Subscription> list = (List<Subscription>) subscriptionRepository.findAll();
+
+        if(list.size() == 0){
+            logger.error("No subscriptions found in the database");
+        }
+        else{
+            logger.info("Getting subscriptions from database");
+        }
+        return list;
     }
 
     @GetMapping("/{id}")
-    public Subscription getSubscription(@PathVariable long id) { return subscriptionRepository.findById(id).get(); }
+    public Subscription getSubscription(@PathVariable long id) {
+
+        Subscription subscription = null;
+        if(subscriptionRepository.findById(id).isPresent()){
+            logger.info("Getting subscription with id {}", id);
+            subscription = subscriptionRepository.findById(id).get();
+        }
+        else{
+            logger.error("Subscription with id {} not found", id);
+        }
+        return subscription;
+    }
 
     @PostMapping("/add")
     public Subscription addSubscription(@RequestBody Subscription subscription) {
-        return subscriptionRepository.save(subscription);
+
+        Subscription newSubscription = subscriptionRepository.save(subscription);
+        logger.info("Saving subscription with id {} to the database", newSubscription.getId());
+        
+        return newSubscription;
     }
 
     @PutMapping("/{id}/update")
     public Subscription updateSubscription(@PathVariable Long id, @RequestBody Subscription subscription) {
+
         subscription.setId(id);
-        return subscriptionRepository.save(subscription);
+        Subscription newSubscription = subscriptionRepository.save(subscription);
+        logger.info("Updating subscription {} , and sending to the database", newSubscription.getId());
+
+        return newSubscription;
     }
+
     @DeleteMapping("/{id}/delete")
     public void deleteSubscription(@PathVariable Long id) {
-        subscriptionRepository.deleteById(id);
+
+        if(subscriptionRepository.findById(id).isPresent()){
+            subscriptionRepository.deleteById(id);
+            logger.info("Deleting subscription with id {}", id);
+        }
+        else{
+            logger.error("Subscription with id {} not found", id);
+        }
 
     }
 }
